+++ tests/src/api/buffer.rs
@@ -158,9 +158,31 @@ fn buf_set_get_del_keymap() {
 
     let keymaps = buf.get_keymap(Mode::Insert).unwrap().collect::<Vec<_>>();
     assert_eq!(1, keymaps.len());
+    assert!(keymaps.iter().all(|keymap| keymap.buffer == Some(buf.clone())));
 
     let res = buf.del_keymap(Mode::Insert, "a");
     assert_eq!(Ok(()), res);
+}
+
+/// Regression test for https://github.com/noib3/nvim-oxi/issues/226
+#[nvim_oxi::test]
+fn buf_set_get_keymap_with_bufnr_more_than_one() {
+    let mut buf = api::create_buf(true, false).unwrap();
+
+    assert!(buf.handle() > 1);
+
+    let opts = SetKeymapOpts::builder()
+        .callback(|_| ())
+        .desc("does nothing")
+        .expr(true)
+        .build();
+
+    let res = buf.set_keymap(Mode::Insert, "a", "", &opts);
+    assert_eq!(Ok(()), res);
+
+    let keymaps = buf.get_keymap(Mode::Insert).unwrap().collect::<Vec<_>>();
+    assert_eq!(1, keymaps.len());
+    assert!(keymaps.iter().all(|keymap| keymap.buffer == Some(buf.clone())));
 }
 
 #[nvim_oxi::test]
@@ -180,6 +202,7 @@ fn buf_set_get_del_nvo_keymap() {
         .unwrap()
         .collect::<Vec<_>>();
     assert_le!(1, keymaps.len());
+    assert!(keymaps.iter().all(|keymap| keymap.buffer == Some(buf.clone())));
 
     let res = buf.del_keymap(Mode::NormalVisualOperator, "a");
     assert_eq!(Ok(()), res);